python/requirements.txt
# ---------------- GPU wheels (CUDA 12.1) ----------------
--extra-index-url https://download.pytorch.org/whl/cu121

torch==2.2.0+cu121
torchvision==0.17.0+cu121          
torchaudio==2.2.0+cu121            
onnxruntime-gpu==1.19.0            

# ---------------- Core ML / vision ----------------------
numpy==1.24.4
scipy==1.10.1
opencv-python==4.8.1.78
insightface==0.7.3
ultralytics==8.3.29

# ---------------- HTTP / API clients --------------------
aiohttp==3.9.5
requests==2.31.0

# ---------------- Speech & I/O --------------------------
edge-tts==6.1.19
aiofiles==24.1.0
python-socketio[asyncio]==5.10.0
randomname==0.2.1

# ---------------- Misc utilities used by the code -------
pandas==2.2.3
tqdm==4.67.0
pillow==11.0.0
scikit-image==0.24.0
scikit-learn==1.6.1
python-dotenv==1.0.0
pydantic[dotenv]==1.10.13

# ---------------- Database / ORM ---------------------------
sqlalchemy>=2.0.0
alembic>=1.13.0
sqids>=0.4.0

# ---------------- Content generation --------------------
openai==1.82.0
huggingface-hub==0.25.2
